Added Ui State for myRecipe and connected it to MyRecipe view models. Added methods in view model for the Ui State

+++ app/src/main/java/com/example/dishpedia/data/MyRecipe.kt
@@ -16,8 +16,8 @@ data class MyRecipe(
     val image: String,
     val title: String = "",
     val summary: String = "",
-    val readyInMinutes: Int = 0,
-    val servings: Int = 0,
+    val readyInMinutes: Double = 0.0,
+    val servings: Double = 0.0,
     val ingredient: ArrayList<String>,
     val instructions: ArrayList<String>
 )
